pyproject.toml
[project]
name = "scout-awards"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=9.0",
    "python-pptx>=0.6",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[tool.setuptools.packages.find]
include = ["scout_awards*"]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"
